pyproject.toml
[project]
name = "agent-ai-playground"
version = "0.7.0"
description = 'This repo contains the source code for my Agentic AI series.'
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "tavily-python",
    "openai",
    "llama-index-llms-openai-like",
    "llama-index",
    "chainlit",
    "rich",
    "pydantic",
    "autogen-ext[chromadb,docker,magentic-one,openai,web-surfer, docker-jupyter-executor]>=0.6.2",
    "autogen-core>=0.6.2",
    "autogen-agentchat>=0.6.2",
    "mcp-server-fetch",
    "mlflow>=3.1",
    "nbclient",
]

[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"